Add a new example website to README.md
Recently, I started working on a hobby project of mine whose purpose is to help & support upcoming artists and I decided to use the Tailwind Nextjs Starter Blog, which was the perfect fit for my case: finding something that is easy to modify, use and deploy. 

As it seems like on the site, I modified a lot on the "core" of Tailwind Nextjs Starter Blog and I also separated two blog parts: the actual blogging feature for my site and the lyrics part: they also generate their own sitemap, tags and RSS files, and the greatest part of it that it was really easy to add, because nowadays, sadly, I don't really have the time to work on a personal project. This is why I'd like to recommend Tailwind Nextjs Starter Blog to everyone. 👍